% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{p1dps}
\alias{p1dps}
\title{List of 143 P1 neurons from Chiang et al 2011 converted to dotprops objects}
\description{
This R list (which has additional class \code{neuronlist}) contains 143
skeletonized \emph{Drosophila} P1 neurons as \code{dotprops} objects.
Original data is due to Chiang et al. 2011, who have generously shared their
raw data at \url{http://flycircuit.tw}. Image registration and further
processing was carried out by Greg Jefferis and Marta Costa.
}
\references{
[1] Chiang A.S., Lin C.Y., Chuang C.C., Chang H.M., Hsieh C.H.,
  Yeh C.W., Shih C.T., Wu J.J., Wang G.T., Chen Y.C., Wu C.C., Chen G.Y.,
  Ching Y.T., Lee P.C., Lin C.Y., Lin H.H., Wu C.C., Hsu H.W., Huang Y.A.,
  Chen J.Y., et al. (2011). Three-dimensional reconstruction of brain-wide
  wiring networks in Drosophila at single-cell resolution. Curr Biol 21 (1),
  1--11.
}

